task/wg-198-Pull-out-pages-react-refactor
