+++ react/src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './components/AppRouter';
+import App from './AppRouter';
 import './index.css';
 import store from './redux/store';
 import { Provider } from 'react-redux';
