+++ react/src/pages/Callback/Callback.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import { MemoryRouter } from 'react-router-dom';
 import { render } from '@testing-library/react';
-import store from '../../../redux/store';
+import store from '../../redux/store';
 import Callback from './Callback';
 
 test('renders callback', async () => {
