+++ react/src/pages/Login/Login.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, waitFor } from '@testing-library/react';
 import Login from './Login';
 import { Provider } from 'react-redux';
-import store from '../../../redux/store';
+import store from '../../redux/store';
 import { MemoryRouter } from 'react-router';
 
 test('renders login', async () => {
