+++ react/src/pages/Logout/Logout.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import Logout from './Logout';
 import { Provider } from 'react-redux';
-import store from '../../../redux/store';
+import store from '../../redux/store';
 import { BrowserRouter } from 'react-router-dom';
 
 test('renders logout', async () => {
