+++ react/src/pages/Logout/Logout.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
-import { logout } from '../../../redux/authSlice';
+import { logout } from '../../redux/authSlice';
 
 function Logout() {
   const dispatch = useDispatch();
