+++ react/src/pages/MainMenu/MainMenu.test.tsx
