+++ react/src/pages/MainMenu/MainMenu.tsx
