+++ react/src/pages/MainMenu/index.ts
